pyproject.toml
[tool.poetry]
name = "news-scraper"
version = "0.1.0"
description = "News aggregation platform with AI-powered analysis"
authors = ["pattersonrptr <pattersonjunior@gmail.com>"]
readme = "README.md"
packages = [{include = "src", from = "backend"}]

[tool.poetry.scripts]
seed-sources = "backend.src.interfaces.cli.seed_sources:main"

[tool.poetry.dependencies]
python = "^3.12"

# Web framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.13.0"
asyncpg = "^0.30.0"        # PostgreSQL async driver
aiosqlite = "^0.20.0"      # SQLite async driver (local dev)

# Task queue
celery = {extras = ["redis"], version = "^5.4.0"}
redis = {extras = ["hiredis"], version = "^5.0.0"}
flower = "^2.0.0"

# Feed collection
feedparser = "^6.0.11"
httpx = "^0.27.0"

# AI providers
google-generativeai = "^0.7.0"   # Gemini
ollama = "^0.2.0"                 # Ollama Python client

# Data validation & config
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-settings = "^2.3.0"

# Utilities
langdetect = "^1.0.9"     # Language detection
python-dotenv = "^1.0.0"
structlog = "^24.1.0"     # Structured logging
tenacity = "^8.3.0"       # Retry / exponential backoff

# Email
aiosmtplib = "^3.0.0"
jinja2 = "^3.1.0"         # HTML email templates

# Auth
python-jose = {extras = ["cryptography"], version = "^3.3.0"}  # JWT
passlib = {extras = ["bcrypt"], version = "^1.7.4"}             # Password hashing
bcrypt = "4.0.1"                                                 # Pinned — passlib 1.7.4 is not compatible with bcrypt 5.x
python-multipart = "^0.0.9"                                     # Form parsing (OAuth2)

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"          # AsyncClient for API tests
respx = "^0.21.0"          # Mock httpx requests
freezegun = "^1.5.0"       # Freeze time in tests
factory-boy = "^3.3.0"     # Test data factories

# Linting & formatting
ruff = "^0.4.4"

# Type checking
mypy = "^1.10.0"
types-redis = "^4.6.0"

# Pre-commit
pre-commit = "^3.7.0"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["backend/src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "ANN", # flake8-annotations
    "RUF", # ruff-specific
]
ignore = [
    "ANN101",  # Missing type annotation for self
    "ANN102",  # Missing type annotation for cls
    "S101",    # Use of assert (allowed in tests)
]

[tool.ruff.lint.per-file-ignores]
"backend/tests/**" = ["S101", "ANN"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
addopts = "--cov=backend/src --cov-report=term-missing --cov-fail-under=30"

[tool.coverage.run]
source = ["backend/src"]
omit = ["backend/tests/*", "backend/src/infrastructure/database/migrations/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
